requirements.txt
mamba-ssm
numpy
pillow
torch
torchvision
timm
tqdm
einops
scikit-learn
fastprogress
pytorch-lightning
fvcore